values_model/src/step1/resolve.py
import os
import json
from typing import List, Dict
from dotenv import load_dotenv

# We are importing the OpenAI library, but the script will fail if run
# because an API key is not configured in this environment.
# import openai

# Load environment variables from a .env file
load_dotenv()

def create_llm_prompt(statement: Dict, all_sentences: List[str]) -> str:
    """
    Creates a detailed, structured prompt for an LLM to correct a flawed extraction.
    """
    idx = statement['sentence_index']
    context_window = 2

    # Assemble the context
    start = max(0, idx - context_window)
    end = min(len(all_sentences), idx + context_window + 1)
    context_lines = []
    for i in range(start, end):
        prefix = "[TARGET SENTENCE]" if i == idx else f"[CONTEXT]"
        context_lines.append(f"{prefix} (index {i}): {all_sentences[i]}")
    context_block = "\n".join(context_lines)

    # The detailed instructions for the LLM, as you designed.
    instructions = '''You are an expert linguist. Your task is to analyze a sentence and correct a flawed concept that was extracted from it. Follow these steps:

    1.  First, analyze the `flawed_extraction` data below. Does the `concept` accurately and precisely capture the idea, action, or entity being evaluated in the `TARGET SENTENCE`?
    2.  If it is not accurate, re-read the `TARGET SENTENCE` and identify the correct concept.
    3.  If the corrected concept contains a pronoun (it, he, she, they, that, etc.), examine the surrounding `CONTEXT` sentences to determine what the pronoun refers to. Add this information to a `references` field.
    4.  If you determine that no clear value judgment is being made in the sentence, return an empty JSON object: {}.
    5.  Return ONLY a single, valid JSON object containing the corrected data. The `concept` must be a direct quote.
    '''

    # Construct the final prompt
    prompt = (
        f"{instructions}\n\n"
        f"CONTEXT BLOCK:\n---\n{context_block}\n---\n\n"
        f"FLAWED EXTRACTION:\n---\n{json.dumps(statement, indent=2)}\n---\n\n"
        f"Return your corrected JSON object below:\n"
    )
    return prompt

def resolve_concepts_with_llm(statements: List[Dict], all_sentences: List[str]) -> List[Dict]:
    """
    Uses a hypothetical LLM call to correct and enrich the extracted value statements.
    """
    api_key = os.getenv("OPENAI_API_KEY")
    if not api_key:
        print("WARNING: OPENAI_API_KEY not found in environment variables.")
        print("         Skipping LLM resolution step. The output will be uncorrected.")
        return statements

    # In a real application, you would initialize the OpenAI client here:
    # client = openai.OpenAI(api_key=api_key)

    corrected_statements = []
    for statement in statements:
        prompt = create_llm_prompt(statement, all_sentences)

        # --- HYPOTHETICAL API CALL ---
        # The following block is what the code would look like.
        # It is commented out because we cannot execute it in this environment.
        # try:
        #     response = client.chat.completions.create(
        #         model="gpt-4-turbo",
        #         messages=[{"role": "system", "content": prompt}],
        #         response_format={"type": "json_object"}
        #     )
        #     corrected_data = json.loads(response.choices[0].message.content)
        #     if corrected_data: # Only append if the LLM didn't return an empty object
        #         corrected_statements.append(corrected_data)
        # except Exception as e:
        #     print(f"An error occurred during the API call: {e}")
        #     # Keep the original statement if the API call fails
        #     corrected_statements.append(statement)

        # For this simulation, since we can't make the API call,
        # we will just return the original statements.
        corrected_statements.append(statement)

    return corrected_statements