requirements.txt
pypdf==5.0.0
python-docx==1.1.2
markdown-it-py==3.0.0
beautifulsoup4==4.12.3
spacy==3.8.7
ftfy==6.2.0
unidecode==1.3.8
orjson==3.10.7
pyarrow==21.0.0
# Sentiment analysis and NLP enhancements
vaderSentiment==3.3.2
textblob==0.17.1
openai==1.108.1
python-dotenv==1.0.1

